+++ packages/apollo-link-batch-http/src/batchHttpLink.ts
@@ -1,6 +1,12 @@
-import { ApolloLink, Operation, FetchResult, Observable } from 'apollo-link';
 import {
-  serializeFetchBody,
+  ApolloLink,
+  Operation,
+  FetchResult,
+  Observable,
+  fromError,
+} from 'apollo-link';
+import {
+  serializeFetchParameter,
   selectURI,
   parseAndCheckHttpResponse,
   checkFetcher,
@@ -102,12 +108,17 @@ export class BatchHttpLink extends ApolloLink {
       const body = optsAndBody.map(({ body }) => body);
       const options = optsAndBody[0].options;
 
+      // There's no spec for using GET with batches.
+      if (options.method === 'GET') {
+        return fromError<FetchResult[]>(
+          new Error('apollo-link-batch-http does not support GET requests'),
+        );
+      }
+
       try {
-        (options as any).body = serializeFetchBody(body);
+        (options as any).body = serializeFetchParameter(body, 'Payload');
       } catch (parseError) {
-        return new Observable<FetchResult[]>(observer => {
-          observer.error(parseError);
-        });
+        return fromError<FetchResult[]>(parseError);
       }
 
       const { controller, signal } = createSignalIfSupported();