+++ packages/apollo-link/src/index.ts
@@ -1,7 +1,12 @@
 export * from './link';
-export { createOperation, makePromise, toPromise, fromPromise } from './linkUtils';
+export {
+  createOperation,
+  makePromise,
+  toPromise,
+  fromPromise,
+  fromError,
+} from './linkUtils';
 export * from './types';
 
-import * as Observable from 'zen-observable';
-
+import Observable from 'zen-observable-ts';
 export { Observable };