+++ packages/zen-observable-ts/rollup.config.js
@@ -1,3 +1,3 @@
 import build from '../../rollup.config';
 
-export default build('schema');
+export default build('zenObservable');